give fullURIIntoRepoAndTag a single function
