+++ app/docker/helpers/imageHelper.js
@@ -16,27 +16,10 @@ function ImageHelperFactory() {
   /**
    *
    * @param {PorImageRegistryModel} registry
-   * @returns {ImageConfig}
-   * @typedef ImageConfig = {
-   *  fromImage: string; // full image URI (including tag)
-   *  repo: string; // image URI without tag
-   *  tag: string;  // image tag
-   * }
    */
   function createImageConfigForContainer(imageModel) {
-    const fromImage = buildImageFullURIFromModel(imageModel);
-    // possible fromImage values:
-    // - registry/image-repo:tag
-    // - image-repo:tag
-    // - registry:port/image-repo:tag
-    // buildImageFullURIFromModel always gives a tag (defaulting to 'latest'), so the tag is always present after the last ':'
-    const parts = fromImage.split(':');
-    const tag = parts.pop() || 'latest';
-    const repo = parts.join(':');
     return {
       fromImage: buildImageFullURIFromModel(imageModel),
-      repo,
-      tag,
     };
   }
