+++ app/react/docker/images/types.ts
@@ -1,3 +1,5 @@
+import { Registry } from '@/react/portainer/registries/types/registry';
+
 import { DockerImageResponse } from './types/response';
 
 type DecoratedDockerImage = {
@@ -6,3 +8,9 @@ type DecoratedDockerImage = {
 
 export type DockerImage = DecoratedDockerImage &
   Omit<DockerImageResponse, keyof DecoratedDockerImage>;
+
+export type ImageModel = {
+  UseRegistry: boolean;
+  Registry?: Registry;
+  Image: string;
+};
